+++ app/Events/AbsenceRequestUpdated.php
@@ -5,11 +5,12 @@ namespace App\Events;
 use App\Models\AbsenceRequest;
 use Illuminate\Broadcasting\Channel;
 use Illuminate\Contracts\Broadcasting\ShouldBroadcast;
+use Illuminate\Foundation\Events\Dispatchable;
 use Illuminate\Queue\SerializesModels;
 
 class AbsenceRequestUpdated implements ShouldBroadcast
 {
-    use SerializesModels;
+    use Dispatchable, SerializesModels;
 
     public function __construct(public AbsenceRequest $absenceRequest)
     {